coin_coding_test/requirements.txt
omegaconf
matplotlib
pillow
numpy
pytorch-lightning==1.2.8
